Refactor source file locations and add unit tests.
Update IO class and tests.

Updated IOTest to have null condition on read input stream.

Add OpenFile tests.

Add more tests for screen.

Added Keyboard tests.

Update QuitActionListener.

Refactor stepping and tracing.

Add ability to step back through code execution.

Consolidate comments.
