--- src/com/chip8java/emulator/listeners/QuitMenuItemActionListener.java
@@ -1,27 +1,0 @@
-/*
- * Copyright (C) 2013-2018 Craig Thomas
- * This project uses an MIT style license - see LICENSE for details.
- */
-package com.chip8java.emulator.listeners;
-
-import javax.swing.*;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-
-/**
- * An ActionListener that will quit the emulator.
- */
-public class QuitMenuItemActionListener implements ActionListener
-{
-    @Override
-    public void actionPerformed(ActionEvent e) {
-        int result = JOptionPane.showConfirmDialog(
-                null,
-                "Are you sure you want to quit?",
-                "Confirm Action",
-                JOptionPane.OK_CANCEL_OPTION);
-        if (result == JOptionPane.OK_OPTION) {
-            System.exit(0);
-        }
-    }
-}
