+++ src/main/java/com/chip8java/emulator/components/EmulatorState.java
@@ -1,0 +1,10 @@
+/*
+ * Copyright (C) 2013-2019 Craig Thomas
+ * This project uses an MIT style license - see LICENSE for details.
+ */
+package com.chip8java.emulator.components;
+
+public enum EmulatorState
+{
+    PAUSED, TRACE, STEP, RUNNING, KILLED
+}
