+++ src/main/java/com/chip8java/emulator/components/ScreenMode.java
