+++ src/main/java/com/chip8java/emulator/listeners/TraceMenuItemListener.java
