+++ src/test/java/com/chip8java/emulator/components/CentralProcessingUnitTest.java
