+++ src/test/resources/test_stream_file.bin
@@ -1,0 +1,1 @@
+This is a test
